requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
SQLAlchemy==2.0.36
python-dotenv==1.0.1
httpx==0.27.2
requests==2.32.3
python-multipart==0.0.20
pdfplumber==0.11.5
python-docx==1.1.2
rapidfuzz==3.14.1
pymupdf==1.26.4
pytesseract==0.3.13
Pillow==11.3.0
streamlit==1.49.1
pandas==2.3.2
